+++ .vscode/settings.json
@@ -6,6 +6,7 @@
     "python.testing.unittestEnabled": false,
     "python.testing.pytestEnabled": true,
     "cSpell.words": [
+        "electoralyze",
         "pytests"
     ]
 }
