+++ README.md
@@ -1,6 +1,5 @@
 # Project
 Electorate x Analyze: Elec-tora-lyze
-<!--  -->
 
 ## Install
