+++ packages/electoralyze/electoralyze/__init__.py
@@ -1,0 +1,3 @@
+from electoralyze.hello import say_hello
+
+__all__ = ["say_hello"]
