+++ packages/electoralyze/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "pixi_py"
+name = "electoralyze"
 description = "Add a short description here"
 authors = [{name = "Andre Medina", email = "AndreMedina@hotmail.com.au"}]
 readme="README.md"
@@ -11,10 +11,10 @@ requires = ["hatchling", "hatch-vcs"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["pixi_py"]
+packages = ["electoralyze"]
 
 [tool.hatch.build.targets.wheel.sources]
-"pixi_py" = "pixi_py"
+"electoralyze" = "electoralyze"
 
 [tool.hatch.version]
 source = "vcs"
