--- packages/pixi_py/pixi_py/__init__.py
@@ -1,3 +1,0 @@
-from pixi_py.hello import say_hello
-
-__all__ = ["say_hello"]
