--- packages/pixi_py_copy/pixi_py_copy/__init__.py
@@ -1,3 +1,0 @@
-from pixi_py_copy.hello import say_hello
-
-__all__ = ["say_hello"]
