+++ packages/ui/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "pixi_py_copy"
+name = "ui"
 description = "Add a short description here"
 authors = [{name = "Andre Medina", email = "AndreMedina@hotmail.com.au"}]
 readme="README.md"
@@ -11,10 +11,10 @@ requires = ["hatchling", "hatch-vcs"]
 build-backend = "hatchling.build"
 
 [tool.hatch.build.targets.wheel]
-packages = ["pixi_py_copy"]
+packages = ["ui"]
 
 [tool.hatch.build.targets.wheel.sources]
-"pixi_py_copy" = "pixi_py_copy"
+"ui" = "ui"
 
 [tool.hatch.version]
 source = "vcs"
